# schemas/openapi.json
{
  "components": {
    "schemas": {
...
      "ClipOut": {
        "additionalProperties": false,
        "properties": {
          "file_status": {
            "$ref": "#/components/schemas/FileStatus"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_reference": {
            "title": "Is Reference",
            "type": "boolean"
          },
          "media": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MediaInfo"
              },
              {
                "type": "null"
              }
            ]
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "path": {
            "title": "Path",
            "type": "string"
          },
          "project_id": {
            "format": "uuid",
            "title": "Project Id",
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/ClipRole"
          },
          "speaker_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Speaker Label"
          },
          "sync": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SyncResult"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "id",
          "project_id",
          "path",
          "name",
          "role",
          "speaker_label",
          "media",
          "sync",
          "file_status",
          "is_reference"
        ],
        "title": "ClipOut",
        "type": "object"
      },
      "ClipRole": {
        "enum": [
          "speaker",
          "wide",
          "broll"
        ],
        "title": "ClipRole",
        "type": "string"
      },
...
      "FileStatus": {
        "enum": [
          "ok",
          "missing",
          "changed",
          "unknown"
        ],
        "title": "FileStatus",
        "type": "string"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MediaInfo": {
        "additionalProperties": false,
        "description": "Facts about a source file, read with ffprobe (Phase 1).",
        "properties": {
          "audio_channels": {
            "default": 0,
            "minimum": 0.0,
            "title": "Audio Channels",
            "type": "integer"
          },
          "audio_codec": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Audio Codec"
          },
          "audio_sample_rate": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Audio Sample Rate"
          },
          "duration_frames": {
            "minimum": 0.0,
            "title": "Duration Frames",
            "type": "integer"
          },
          "fps": {
            "$ref": "#/components/schemas/Rational"
          },
          "height": {
            "exclusiveMinimum": 0.0,
            "title": "Height",
            "type": "integer"
          },
          "is_vfr": {
            "title": "Is Vfr",
            "type": "boolean"
          },
          "video_codec": {
            "title": "Video Codec",
            "type": "string"
          },
          "width": {
            "exclusiveMinimum": 0.0,
            "title": "Width",
            "type": "integer"
          }
        },
        "required": [
          "fps",
          "is_vfr",
          "duration_frames",
          "width",
          "height",
          "video_codec",
          "audio_codec",
          "audio_sample_rate",
          "audio_channels"
        ],
        "title": "MediaInfo",
        "type": "object"
      },
      "OutputSettings": {
        "additionalProperties": false,
        "properties": {
          "fps": {
            "$ref": "#/components/schemas/Rational"
          },
          "height": {
            "exclusiveMinimum": 0.0,
            "title": "Height",
            "type": "integer"
          },
          "width": {
            "exclusiveMinimum": 0.0,
            "title": "Width",
            "type": "integer"
          }
        },
        "required": [
          "fps",
          "width",
          "height"
        ],
        "title": "OutputSettings",
        "type": "object"
      },
      "Preset": {
        "enum": [
          "calm",
          "balanced",
          "dynamic"
        ],
        "title": "Preset",
        "type": "string"
      },
      "ProjectCreate": {
        "additionalProperties": false,
        "properties": {
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "output": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OutputSettings"
              },
              {
                "type": "null"
              }
            ],
            "description": "Omit to follow the reference clip's frame rate and size"
          },
          "preset": {
            "$ref": "#/components/schemas/Preset",
            "default": "balanced"
          }
        },
        "required": [
          "name"
        ],
        "title": "ProjectCreate",
        "type": "object"
      },
      "ProjectOut": {
        "additionalProperties": false,
        "properties": {
          "clips": {
            "items": {
              "$ref": "#/components/schemas/ClipOut"
            },
            "title": "Clips",
            "type": "array"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "cutlist_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cutlist Version"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "output": {
            "$ref": "#/components/schemas/OutputSettings"
          },
          "output_custom": {
            "title": "Output Custom",
            "type": "boolean"
          },
          "preset": {
            "$ref": "#/components/schemas/Preset"
          },
          "reference_clip_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reference Clip Id"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "preset",
          "output",
          "output_custom",
          "reference_clip_id",
          "clips",
          "cutlist_version",
          "created_at",
          "updated_at"
        ],
        "title": "ProjectOut",
        "type": "object"
      },
      "ProjectSummary": {
        "additionalProperties": false,
        "properties": {
          "clip_count": {
            "title": "Clip Count",
            "type": "integer"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "cutlist_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cutlist Version"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "preset": {
            "$ref": "#/components/schemas/Preset"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "preset",
          "clip_count",
          "cutlist_version",
          "created_at",
          "updated_at"
        ],
        "title": "ProjectSummary",
        "type": "object"
      },
      "ProjectUpdate": {
        "additionalProperties": false,
        "properties": {
          "name": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "output": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OutputSettings"
              },
              {
                "type": "null"
              }
            ]
          },
          "preset": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Preset"
              },
              {
                "type": "null"
              }
            ]
          },
          "reference_clip_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reference Clip Id"
          }
        },
        "title": "ProjectUpdate",
        "type": "object"
      },
      "Rational": {
        "additionalProperties": false,
        "description": "An exact positive rational number, e.g. a frame rate of 30000/1001.",
        "properties": {
          "den": {
            "exclusiveMinimum": 0.0,
            "title": "Den",
            "type": "integer"
          },
          "num": {
            "exclusiveMinimum": 0.0,
            "title": "Num",
            "type": "integer"
          }
        },
        "required": [
          "num",
          "den"
        ],
        "title": "Rational",
        "type": "object"
      },
...
      "SyncResult": {
        "additionalProperties": false,
        "description": "How a clip lines up with the project's reference clip.\n\nSign convention (shared with ``benchmark.GroundTruth``):\n    ``offset_samples`` = position of an event in THIS clip\n                         minus position of the same event in the REFERENCE clip.\n    Positive  -> this clip started recording EARLIER than the reference.\n    Negative  -> this clip started recording LATER than the reference.\n\n``drift_ppm``: how much faster this clip's clock runs than the reference,\nin parts per million. A value of +100 means the clip accumulates 0.36 s\nof extra duration per hour. This is a model parameter, not a timeline\nposition, so a float is acceptable here.",
        "properties": {
          "confidence": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "type": "number"
          },
          "drift_ppm": {
            "default": 0.0,
            "title": "Drift Ppm",
            "type": "number"
          },
          "offset_samples": {
            "title": "Offset Samples",
            "type": "integer"
          },
          "reference_clip_id": {
            "format": "uuid",
            "title": "Reference Clip Id",
            "type": "string"
          },
          "sample_rate": {
            "exclusiveMinimum": 0.0,
            "title": "Sample Rate",
            "type": "integer"
          }
        },
        "required": [
          "reference_clip_id",
          "offset_samples",
          "sample_rate",
          "drift_ppm",
          "confidence"
        ],
        "title": "SyncResult",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Local backend: projects, clips, jobs (sync, auto-edit, render).",
    "title": "Multicam Studio API",
    "version": "0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/projects": {
      "get": {
        "operationId": "list_projects_api_projects_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProjectSummary"
                  },
                  "title": "Response List Projects Api Projects Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Projects",
        "tags": [
          "projects"
        ]
      },
      "post": {
        "operationId": "create_project_api_projects_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Project",
        "tags": [
          "projects"
        ]
      }
    },
    "/api/projects/{project_id}": {
      "delete": {
        "description": "Deletes the project and its caches/renders in the data folder. Your media is untouched.",
        "operationId": "delete_project_api_projects__project_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Project",
        "tags": [
          "projects"
        ]
      },
      "get": {
        "operationId": "get_project_api_projects__project_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Project",
        "tags": [
          "projects"
        ]
      },
      "patch": {
        "operationId": "update_project_api_projects__project_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Project",
        "tags": [
          "projects"
        ]
      }
    },
...
  }
}